requirements.txt
ollama==0.1.0
requests==2.31.0
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
pypdf==4.0.1
pdfplumber==0.10.3
python-docx==0.8.11
pandas
pyyaml==6.0.1
pytest==7.4.3
python-dotenv==1.0.1
openpyxl==3.11.0
spacy==3.7.2

